README.md: pip3 more likely correct than pip
The scripts say python3, so assuming they find the right python3
executable, you're more likely to get the right pip for them by typing
'pip3'.

A separarate issue not addressed here: The way to invoke Python needs
to be a configuration parameter. The correct name to get Python will
vary, and systems that don't ship Python 2 anymore don't necessarily
have a python3 (just calling it python instead).

Also, --user is likely needed if not in a container or if running as root.