configs: ts7680: initial commit
Minimal and headless configurations

Currently uses a bit of a hack of including the U-Boot binary for
TS-7670 so the final output image can be made without genimage
erroring. The TS-7680 doesn't need an external U-Boot binary, but
in order to not make more genimage configs, we just add an unrelated
U-Boot binary to the final image that is benign, and takes up no
additional space.

Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>